feature: zlog_init_from_string()
Added a new init function which takes the configuration as a
string instead of a path to a configuration file to be read.
This allows the configuration to be embed in the application to
avoid yet another configuration file to be managed (and
potentially broken after deployment.)

The embedded configuration string processing is the same as an
external zlog configuration file, with the exception that
line-contination characters are not supported.